Split fixture file
