+++ apps/sentinel/test/support/monitoring/beam_vm.ex
@@ -1,9 +1,8 @@
 defmodule Sentinel.Fixture.Monitoring.BeamVm do
   @moduledoc """
-  This module will handle Beam Vm Fixtures
+  This module will provide Beam Vm Fixtures
   """
 
-  alias Host.Memory
   alias Sentinel.Monitoring.BeamVm
 
   def update_app_message(source_node, node, attrs) do
@@ -27,15 +26,5 @@ defmodule Sentinel.Fixture.Monitoring.BeamVm do
     }
 
     {:beam_vm_update_statistics, message}
-  end
-
-  def update_sys_info_message(source_node, memory_free, memory_total) do
-    message = %Memory{
-      source_node: source_node,
-      memory_free: memory_free,
-      memory_total: memory_total
-    }
-
-    {:update_system_info, message}
   end
 end
